+++ modules/utils-compose/src/commonMain/kotlin/com/thekeeperofpie/artistalleydatabase/utils_compose/ShareHandler.kt
@@ -1,0 +1,9 @@
+package com.thekeeperofpie.artistalleydatabase.utils_compose
+
+import androidx.compose.runtime.compositionLocalOf
+
+expect class ShareHandler {
+    fun shareUrl(title: String?, url: String)
+}
+
+val LocalShareHandler = compositionLocalOf<ShareHandler?> { null }